+++ packages/core/src/utils/formatQuery/__tests__/formatQuery.SQL.test.ts
@@ -18,6 +18,7 @@ import { defaultRuleProcessorParameterized } from '../defaultRuleProcessorParame
 import { defaultRuleProcessorSQL } from '../defaultRuleProcessorSQL';
 import { formatQuery } from '../formatQuery';
 import {
+  getMuteTestData,
   getValidationTestData,
   query,
   queryAllOperators,
@@ -449,6 +450,83 @@ describe('independent combinators', () => {
         'parameterized_named'
       )
     ).toEqual({ sql: `(field = :field_1 and (1 = 1))`, params: { field_1: '' } });
+  });
+});
+
+describe('muted', () => {
+  describe('sql', () => {
+    const muteTestResults: Record<string, string> = {
+      'should mute sql': '(1 = 1)',
+      'should mute sql even if fields are valid': '(1 = 1)',
+      'should mute sql rule specified by validationMap': `(field2 = '')`,
+      'should mute sql outermost group': '(1 = 1)',
+      'should mute sql inner group': '((1 = 1))',
+      'should mute sql following combinator of first rule': `(field2 = '' or field3 = '')`,
+      'should mute sql preceding combinator of non-first rule': `(field = '' or field3 = '')`,
+    };
+
+    for (const vtd of getMuteTestData('sql')) {
+      it(vtd.title, () => {
+        expect(formatQuery(vtd.query, vtd.options)).toBe(muteTestResults[vtd.title]);
+      });
+    }
+  });
+
+  describe('parameterized', () => {
+    const muteTestResults: Record<string, ParameterizedSQL> = {
+      'should mute parameterized': { sql: '(1 = 1)', params: [] },
+      'should mute parameterized even if fields are valid': { sql: '(1 = 1)', params: [] },
+      'should mute parameterized rule specified by validationMap': {
+        sql: '(field2 = ?)',
+        params: [''],
+      },
+      'should mute parameterized outermost group': { sql: '(1 = 1)', params: [] },
+      'should mute parameterized inner group': { sql: '((1 = 1))', params: [] },
+      'should mute parameterized following combinator of first rule': {
+        sql: '(field2 = ? or field3 = ?)',
+        params: ['', ''],
+      },
+      'should mute parameterized preceding combinator of non-first rule': {
+        sql: '(field = ? or field3 = ?)',
+        params: ['', ''],
+      },
+    };
+
+    for (const vtd of getMuteTestData('parameterized')) {
+      it(vtd.title, () => {
+        expect(formatQuery(vtd.query, vtd.options)).toEqual(muteTestResults[vtd.title]);
+      });
+    }
+  });
+
+  describe('parameterized_named', () => {
+    const muteTestResults: Record<string, ParameterizedNamedSQL> = {
+      'should mute parameterized_named': { sql: '(1 = 1)', params: {} },
+      'should mute parameterized_named even if fields are valid': {
+        sql: '(1 = 1)',
+        params: {},
+      },
+      'should mute parameterized_named rule specified by validationMap': {
+        sql: '(field2 = :field2_1)',
+        params: { field2_1: '' },
+      },
+      'should mute parameterized_named outermost group': { sql: '(1 = 1)', params: {} },
+      'should mute parameterized_named inner group': { sql: '((1 = 1))', params: {} },
+      'should mute parameterized_named following combinator of first rule': {
+        sql: '(field2 = :field2_1 or field3 = :field3_1)',
+        params: { field2_1: '', field3_1: '' },
+      },
+      'should mute parameterized_named preceding combinator of non-first rule': {
+        sql: '(field = :field_1 or field3 = :field3_1)',
+        params: { field_1: '', field3_1: '' },
+      },
+    };
+
+    for (const vtd of getMuteTestData('parameterized_named')) {
+      it(vtd.title, () => {
+        expect(formatQuery(vtd.query, vtd.options)).toEqual(muteTestResults[vtd.title]);
+      });
+    }
   });
 });